🏗️ Move pages related to organization feature into new dedicated folder
