+++ src/features/organization/pages/ActivityFeed/style.scss
@@ -1,5 +1,5 @@
-@use '../../styles/settings/breakpoints' as b;
-@use '../../styles/settings/colors' as v;
+@use '../../../../styles/settings/breakpoints' as b;
+@use '../../../../styles/settings/colors' as v;
 
 .c-box-main {
     flex-grow: 1;
