+++ src/features/organization/pages/OrganizationCreation/index.jsx
@@ -1,5 +1,5 @@
-import { OrganizationForm } from '../../components/Forms/OrganizationForm'
-import { DoublePageLayout } from '../../layouts/DoublePageLayout'
+import { OrganizationForm } from '../../../../components/Forms/OrganizationForm'
+import { DoublePageLayout } from '../../../../layouts/DoublePageLayout'
 
 import './style.scss'
