+++ src/features/organization/pages/OrganizationCreation/style.scss
@@ -1,0 +1,1 @@
+@use '../../../../styles/settings/colors' as v;
