--- src/pages/OrganizationCreation/style.scss
@@ -1,1 +1,0 @@
-@use '../../styles/settings/colors' as v;
