+++ src/routes/Router.jsx
@@ -3,12 +3,12 @@ import { useSelector } from 'react-redux'
 import { getUserOrganizationId } from '../features/user/store/userSelectors'
 import { UserProfile } from '../features/user/pages/UserProfile'
 import { Home } from '../pages/Home'
-import { OrganizationCreation } from '../pages/OrganizationCreation'
+import { OrganizationCreation } from '../features/organization/pages/OrganizationCreation'
 import { Administration } from '../pages/Administration'
 import { ProfileSettings } from '../features/user/pages/ProfileSettings'
 import { Contact } from '../pages/Contact'
 import { SignUp } from '../features/user/pages/SignUp'
-import { ActivityFeed } from '../pages/ActivityFeed'
+import { ActivityFeed } from '../features/organization/pages/ActivityFeed'
 import { AdminRoute } from './ConditionalRoute/AdminRoute'
 import { GuestRoute } from './ConditionalRoute/GuestRoute'
 import { AuthenticatedRoute } from './ConditionalRoute/AuthenticatedRoute'
